requirements.txt
python-telegram-bot==22.3
apscheduler==3.11.0
pytz
tzlocal
watchgod==0.8.2
python-dotenv==1.0.0